[skip ci] tests/test_units/test_run_local_ssb_selection.py -> tests/test_units/test_path_utils.py
